requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    make up
#
bcrypt==3.1.2             # via -r requirements.in
blinker==1.4              # via -r requirements.in
bravado-core==5.16.1      # via pyramid-swagger
certifi==2019.11.28       # via requests
cffi==1.14.0              # via bcrypt
chardet==3.0.4            # via requests
dnspython==1.16.0         # via -r requirements.in
getconf==1.9.0            # via -r requirements.in
hupper==1.10.2            # via pyramid
idna==2.9                 # via requests
jsonref==0.2              # via bravado-core
jsonschema[format]==2.6.0  # via -r requirements.in, bravado-core, pyramid-swagger, swagger-spec-validator
mongoengine==0.19.1       # via -r requirements.in
msgpack==1.0.0            # via bravado-core
paste==3.4.0              # via pastescript
pastedeploy==2.1.0        # via pastescript, plaster-pastedeploy, pyramid
pastescript==2.0.2        # via -r requirements.in
plaster-pastedeploy==0.7  # via pyramid
plaster==1.0              # via plaster-pastedeploy, pyramid
pycparser==2.20           # via cffi
pyjwt==1.4.2              # via -r requirements.in
pymongo==3.10.1           # via mongoengine
pyramid-swagger==2.7.0    # via -r requirements.in
pyramid==1.9.1            # via -r requirements.in, pyramid-swagger
python-dateutil==2.8.1    # via bravado-core
pytz==2019.3              # via bravado-core
pyyaml==5.3               # via bravado-core, swagger-spec-validator
repoze.lru==0.7           # via pyramid
requests==2.23.0          # via -r requirements.in
rfc3987==1.3.8            # via jsonschema
simplejson==3.17.0        # via bravado-core, pyramid-swagger
six==1.14.0               # via bcrypt, bravado-core, mongoengine, paste, pastescript, python-dateutil, swagger-spec-validator
strict-rfc3339==0.7       # via jsonschema
swagger-spec-validator==2.3.1  # via -r requirements.in, bravado-core
translationstring==1.3    # via pyramid
urllib3==1.25.8           # via requests
uwsgi==2.0.18             # via -r requirements.in
venusian==3.0.0           # via pyramid
webcolors==1.11.1         # via jsonschema
webob==1.8.6              # via pyramid
zope.deprecation==4.4.0   # via pyramid
zope.interface==4.7.2     # via pyramid

# The following packages are considered to be unsafe in a requirements file:
# setuptools
